// src/enums/vm/VMOpcode.ts
export enum VMOpcode {
    Nop = 0,
    Const = 1,
    Return = 2,
    Throw = 3,
    Pop = 4,
    Dup = 5,
    GetLocal = 6,
    SetLocal = 7,
    InitLocal = 8,
    GetCapture = 9,
    SetCapture = 10,
    GetGlobal = 11,
    SetGlobal = 12,
    DeleteGlobal = 13,
    TypeofGlobal = 14,
    This = 15,
    Arguments = 16,
    NewTarget = 17,
    Unary = 18,
    Binary = 19,
    UpdateLocal = 20,
    GetProperty = 21,
    SetProperty = 22,
    DeleteProperty = 23,
    GetPrivate = 24,
    SetPrivate = 25,
    MakeArray = 26,
    ArrayAppend = 27,
    ArraySpread = 28,
    MakeObject = 29,
    DefineProperty = 30,
    ObjectSpread = 31,
    Jump = 32,
    JumpIfTrue = 33,
    JumpIfFalse = 34,
    JumpIfNullish = 35,
    Call = 36,
    CallMethod = 37,
    CallContext = 38,
    Construct = 39,
    MakeClosure = 40,
    IterOpen = 41,
    IterNext = 42,
    IterClose = 43,
    AsyncIterOpen = 44,
    AsyncIterNext = 45,
    Await = 46,
    Yield = 47,
    YieldStar = 48,
    MakeClass = 49,
    SuperGet = 50,
    SuperSet = 51,
    SuperCall = 52,
    PrivateOp = 53,
    DynamicImport = 54,
    EnterWith = 55,
    ExitWith = 56,
    ResolveName = 57,
    SetName = 58,
    EvalThunk = 59,
    Debugger = 60,
    End = 61,
    ImportMeta = 62,
    EnumKeys = 63,
    LocalBinaryConst = 64,
    PropertyCall = 65,
    UpdateReference = 66,
    RegisterMove = 511,
    FirstDecoy = 512
}

export const VM_MACRO_OPCODE_MIN: number = 64;
export const VM_MACRO_OPCODE_MAX: number = 127;
export const VM_REGISTER_OPCODE_OFFSET: number = 256;